Implement changing background on 2D spaceviews
