+++ crates/re_space_view_spatial/src/space_view_2d.rs
@@ -240,14 +240,17 @@ impl SpaceViewClass for SpatialSpaceView2D {
         ui: &mut egui::Ui,
         state: &mut dyn SpaceViewState,
         _space_origin: &EntityPath,
-        _space_view_id: SpaceViewId,
+        space_view_id: SpaceViewId,
         _root_entity_properties: &mut EntityProperties,
     ) -> Result<(), SpaceViewSystemExecutionError> {
         let state = state.downcast_mut::<SpatialSpaceViewState>()?;
         ctx.re_ui
             .selection_grid(ui, "spatial_settings_ui")
             .show(ui, |ui| {
-                state.default_size_ui(ctx, ui);
+                state.default_sizes_ui(ctx, ui);
+
+                crate::ui::background_ui(ctx, space_view_id, ui);
+
                 state.bounding_box_ui(ctx, ui, SpatialSpaceViewKind::TwoD);
 
                 {
