+++ rerun_py/rerun_sdk/rerun/blueprint/views/spatial2d_view.py
@@ -6,8 +6,10 @@ from __future__ import annotations
 __all__ = ["Spatial2DView"]
 
 
+from ... import datatypes
 from ..._baseclasses import AsComponents
 from ...datatypes import EntityPathLike, Utf8Like
+from .. import archetypes as blueprint_archetypes
 from .. import components as blueprint_components
 from ..api import SpaceView, SpaceViewContentsLike
 
@@ -22,6 +24,10 @@ class Spatial2DView(SpaceView):
         contents: SpaceViewContentsLike = "$origin/**",
         name: Utf8Like | None = None,
         visible: blueprint_components.VisibleLike | None = None,
+        background: blueprint_archetypes.Background
+        | datatypes.Rgba32Like
+        | blueprint_components.BackgroundKindLike
+        | None = None,
     ) -> None:
         """
         Construct a blueprint for a new Spatial2DView view.
@@ -41,10 +47,17 @@ class Spatial2DView(SpaceView):
             Whether this view is visible.
 
             Defaults to true if not specified.
+        background:
+            Configuration for the background of the space view.
 
         """
 
         properties: dict[str, AsComponents] = {}
+        if background is not None:
+            if not isinstance(background, blueprint_archetypes.Background):
+                background = blueprint_archetypes.Background(background)
+            properties["Background"] = background
+
         super().__init__(
             class_identifier="2D", origin=origin, contents=contents, name=name, visible=visible, properties=properties
         )
